refactor: move the palette helpers to ai/utils
They sit next to normalize_vegalite_spec now, so anything else that wants the
shared palette can reach it without importing the agent.
